Move the `ModelMetadataTrait` to the correct namespace
