+++ core-bundle/src/File/ModelMetadataTrait.php
@@ -12,52 +12,15 @@ declare(strict_types=1);
 
 namespace Contao\CoreBundle\File;
 
-use Contao\FilesModel;
-use Contao\System;
-use Contao\Validator;
+use Contao\Model\MetadataTrait;
+
+trigger_deprecation('contao/core-bundle', '5.3', 'Using "Contao\CoreBundle\File\ModelMetadataTrait" has been deprecated and will no longer work in Contao 6. Use "Contao\Model\MetadataTrait" instead.');
 
 /**
- * @property string $overwriteMeta
- *
- * @method array row()
+ * @deprecated Deprecated since Contao 5.3, to be removed in Contao 6;
+ *             use Contao\Model\MetadataTrait instead.
  */
 trait ModelMetadataTrait
 {
-    /**
-     * Get the default metadata or null if not applicable.
-     */
-    public function getOverwriteMetadata(): Metadata|null
-    {
-        // Ignore if "overwriteMeta" is not set
-        if (!$this->overwriteMeta) {
-            return null;
-        }
-
-        $data = $this->row();
-
-        // Normalize keys
-        if (isset($data['imageTitle'])) {
-            $data[Metadata::VALUE_TITLE] = $data['imageTitle'];
-        }
-
-        if (isset($data['imageUrl'])) {
-            $url = $data['imageUrl'];
-
-            if (Validator::isRelativeUrl($url)) {
-                $url = System::getContainer()->get('contao.assets.files_context')->getStaticUrl().$url;
-            }
-
-            $data[Metadata::VALUE_URL] = $url;
-        }
-
-        unset($data['imageTitle'], $data['imageUrl']);
-
-        // Make sure we resolve insert tags pointing to files
-        if (isset($data[Metadata::VALUE_URL])) {
-            $data[Metadata::VALUE_URL] = System::getContainer()->get('contao.insert_tag.parser')->replaceInline($data[Metadata::VALUE_URL] ?? '');
-        }
-
-        // Strip superfluous fields by intersecting with tl_files.meta.eval.metaFields
-        return new Metadata(array_intersect_key($data, array_flip(FilesModel::getMetaFields())));
-    }
+    use MetadataTrait;
 }
