+++ faq-bundle/contao/models/FaqModel.php
@@ -10,8 +10,8 @@
 
 namespace Contao;
 
-use Contao\CoreBundle\File\ModelMetadataTrait;
 use Contao\Model\Collection;
+use Contao\Model\MetadataTrait;
 
 /**
  * Reads and writes FAQs
@@ -127,7 +127,7 @@ use Contao\Model\Collection;
  */
 class FaqModel extends Model
 {
-	use ModelMetadataTrait;
+	use MetadataTrait;
 
 	/**
 	 * Table name
